Refactor the Variants namespace, allowing classes to share the exported functions for DRY's sake. Also introduce VariantComparer class to implement the comparing of 2 variants.
