+++ Rubberduck.VBEEditor/Variants/VariantConverter.cs
@@ -4,56 +4,6 @@ using System.Runtime.InteropServices;
 
 namespace Rubberduck.VBEditor.Variants
 {
-    public enum VARENUM
-    {
-        VT_EMPTY = 0x0000,
-        VT_NULL = 0x0001,
-        VT_I2 = 0x0002,
-        VT_I4 = 0x0003,
-        VT_R4 = 0x0004,
-        VT_R8 = 0x0005,
-        VT_CY = 0x0006,
-        VT_DATE = 0x0007,
-        VT_BSTR = 0x0008,
-        VT_DISPATCH = 0x0009,
-        VT_ERROR = 0x000A,
-        VT_BOOL = 0x000B,
-        VT_VARIANT = 0x000C,
-        VT_UNKNOWN = 0x000D,
-        VT_DECIMAL = 0x000E,
-        VT_I1 = 0x0010,
-        VT_UI1 = 0x0011,
-        VT_UI2 = 0x0012,
-        VT_UI4 = 0x0013,
-        VT_I8 = 0x0014,
-        VT_UI8 = 0x0015,
-        VT_INT = 0x0016,
-        VT_UINT = 0x0017,
-        VT_VOID = 0x0018,
-        VT_HRESULT = 0x0019,
-        VT_PTR = 0x001A,
-        VT_SAFEARRAY = 0x001B,
-        VT_CARRAY = 0x001C,
-        VT_USERDEFINED = 0x001D,
-        VT_LPSTR = 0x001E,
-        VT_LPWSTR = 0x001F,
-        VT_RECORD = 0x0024,
-        VT_INT_PTR = 0x0025,
-        VT_UINT_PTR = 0x0026,
-        VT_ARRAY = 0x2000,
-        VT_BYREF = 0x4000
-    }
-
-    [Flags]
-    public enum VariantConversionFlags : ushort
-    {
-        NO_FLAGS = 0x00,
-        VARIANT_NOVALUEPROP = 0x01,     //Prevents the function from attempting to coerce an object to a fundamental type by getting the Value property. Applications should set this flag only if necessary, because it makes their behavior inconsistent with other applications.
-        VARIANT_ALPHABOOL = 0x02,       //Converts a VT_BOOL value to a string containing either "True" or "False".
-        VARIANT_NOUSEROVERRIDE = 0x04,  //For conversions to or from VT_BSTR, passes LOCALE_NOUSEROVERRIDE to the core coercion routines.
-        VARIANT_LOCALBOOL = 0x08        //For conversions from VT_BOOL to VT_BSTR and back, uses the language specified by the locale in use on the local computer. 
-    }
-
     /// <summary>
     /// Handles variant conversions, enabling us to have same implicit conversion behaviors within
     /// .NET as we can observe it from VBA/VB6.
@@ -67,40 +17,18 @@ namespace Rubberduck.VBEditor.Variants
     /// </remarks>
     public static class VariantConverter
     {
-        private const string dllName = "oleaut32.dll";
-
-        // HRESULT VariantChangeType(
-        //   VARIANTARG       *pvargDest,
-        //   const VARIANTARG *pvarSrc,
-        //   USHORT           wFlags,
-        //   VARTYPE          vt
-        // );
-        [DllImport(dllName, EntryPoint = "VariantChangeType", CharSet = CharSet.Auto, SetLastError = true, PreserveSig = true)]
-        private static extern int VariantChangeType(ref object pvargDest, ref object pvarSrc, VariantConversionFlags wFlags, VARENUM vt);
-
-        // HRESULT VariantChangeTypeEx(
-        //   VARIANTARG        *pvargDest,
-        //   const VARIANTARG  *pvarSrc,
-        //   LCID              lcid,
-        //   USHORT            wFlags,
-        //   VARTYPE           vt
-        // );
-        [DllImport(dllName, EntryPoint = "VariantChangeTypeEx", CharSet = CharSet.Auto, SetLastError = true, PreserveSig = true)]
-        private static extern int VariantChangeTypeEx(ref object pvargDest, ref object pvarSrc, int lcid, VariantConversionFlags wFlags, VARENUM vt);
-
         public static object ChangeType(object value, VARENUM vt)
         {
             return ChangeType(value, vt, null);
         }
 
-        private static bool HRESULT_FAILED(int hr) => hr < 0;
         public static object ChangeType(object value, VARENUM vt, CultureInfo cultureInfo)
         {
             object result = null;
             var hr = cultureInfo == null
-                ? VariantChangeType(ref result, ref value, VariantConversionFlags.NO_FLAGS, vt)
-                : VariantChangeTypeEx(ref result, ref value, cultureInfo.LCID, VariantConversionFlags.NO_FLAGS, vt);
-            if (HRESULT_FAILED(hr))
+                ? VariantNativeMethods.VariantChangeType(ref result, ref value, VariantConversionFlags.NO_FLAGS, vt)
+                : VariantNativeMethods.VariantChangeTypeEx(ref result, ref value, cultureInfo.LCID, VariantConversionFlags.NO_FLAGS, vt);
+            if (HResult.Failed(hr))
             {
                 throw Marshal.GetExceptionForHR(hr);
             }
